gate auth

+++ app/Http/Controllers/QuestionController.php
@@ -62,7 +62,11 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         return view("Questions.edit", compact('question'));
+
     }
 
     /**
@@ -74,6 +78,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->update($request-> only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question has been updated'); 
     }
@@ -86,6 +93,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted'); 
     }
